seperating component todoform
