+++ src/views/Todos.vue
@@ -2,13 +2,9 @@
   <header>
     <h3>{{ t("TODO list") }}</h3>
   </header>
-  <form @submit.prevent="handeleSubmit" class="addTodo">
-    <input v-model="newTask" placeholder="Add new Todo" />
-    <button @click="handeleSubmit" :class="tem">{{ t("Add") }}</button>
-  </form>
-
+  <TodoForm @add-Task="addTodo" />
   <div v-for="task of tasks" :key="task.id" class="task-list">
-    <TodoDetail :task="task" @delete-Task="deleteTask" />
+    <TodoDetail :task="task" @delete-Task="deleteTodo" />
   </div>
 </template>
 
@@ -16,9 +12,7 @@
 import { ref } from "vue";
 import { useI18n } from "vue-i18n";
 import TodoDetail from "../component/Todos/TodoDetail.vue";
-
-let tem: string | null;
-tem = localStorage.getItem("thems");
+import TodoForm from "../component/Todos/TodoForm.vue";
 
 const { t } = useI18n({
   inheritLocale: true,
@@ -43,19 +37,15 @@ let tasks = ref([
     id: 4,
   },
 ]);
-const newTask = ref("");
 
-function handeleSubmit() {
-  if (newTask.value.length > 0) {
-    tasks.value.push({
-      title: newTask.value,
-      id: Math.floor(Math.random() * 10000),
-    });
-    newTask.value = "";
-  }
+function addTodo(payload: string) {
+  tasks.value.push({
+    title: payload,
+    id: Math.floor(Math.random() * 10000),
+  });
 }
 
-function deleteTask(payload: task) {
+function deleteTodo(payload: task) {
   tasks.value = tasks.value.filter((task) => payload.id !== task.id);
 }
 </script>
@@ -76,33 +66,6 @@ header h1 {
   color: #777;
   transform: rotate(2deg);
 }
-form {
-  margin: 0 auto;
-  display: grid;
-  grid-template-columns: 3fr 1fr;
-  gap: 10px;
-}
-form button {
-  background: #20b4e5;
-  padding: 10px;
-  border-radius: 10px;
-  border: none;
-  cursor: pointer;
-}
-form input {
-  border: 0;
-  padding: 10px;
-  border-radius: 6px;
-  color: #555;
-  font-size: 1em;
-}
-.addTodo {
-  display: inline-flex;
-}
-.addTodo input {
-  border-radius: 10px;
-}
-
 .task-list {
   max-width: 640px;
   margin: 20px auto;
